Implement creator checks so only creators of resources can modify/delete them

+++ src/main/java/com/mrumstajn/gamedevforum/service/command/impl/PostCommandServiceImpl.java
@@ -3,16 +3,20 @@ package com.mrumstajn.gamedevforum.service.command.impl;
 import com.mrumstajn.gamedevforum.dto.request.CreatePostRequest;
 import com.mrumstajn.gamedevforum.dto.request.EditPostRequest;
 import com.mrumstajn.gamedevforum.entity.Post;
+import com.mrumstajn.gamedevforum.exception.UnauthorizedActionException;
 import com.mrumstajn.gamedevforum.repository.PostRepository;
 import com.mrumstajn.gamedevforum.service.command.PostCommandService;
+import com.mrumstajn.gamedevforum.service.command.UserPostReactionCommandService;
 import com.mrumstajn.gamedevforum.service.query.ForumUserQueryService;
 import com.mrumstajn.gamedevforum.service.query.PostQueryService;
+import com.mrumstajn.gamedevforum.util.UserUtil;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
+import java.util.Objects;
 
 @Service
 @Transactional
@@ -21,6 +25,8 @@ public class PostCommandServiceImpl implements PostCommandService {
     private final PostRepository postRepository;
 
     private final PostQueryService postQueryService;
+
+    private final UserPostReactionCommandService userPostReactionCommandService;
 
     private final ForumUserQueryService forumUserQueryService;
 
@@ -38,6 +44,11 @@ public class PostCommandServiceImpl implements PostCommandService {
     @Override
     public Post edit(Long id, EditPostRequest request) {
         Post existingPost = postQueryService.getById(id);
+
+        if (!isCurrentUserPostOwner(existingPost)){
+            throw new UnauthorizedActionException("User is not the creator of this post");
+        }
+
         modelMapper.map(request, existingPost);
 
         return postRepository.save(existingPost);
@@ -47,6 +58,15 @@ public class PostCommandServiceImpl implements PostCommandService {
     public void delete(Long id) {
         Post existingPost = postQueryService.getById(id);
 
+        if (!isCurrentUserPostOwner(existingPost)){
+            throw new UnauthorizedActionException("User is not the creator of this post");
+        }
+
+        userPostReactionCommandService.deleteAllByPostId(existingPost.getId());
         postRepository.delete(existingPost);
+    }
+
+    private boolean isCurrentUserPostOwner(Post post){
+        return Objects.equals(UserUtil.getCurrentUser().getId(), post.getAuthor().getId());
     }
 }
